+++ src/test/java/com/idilia/samples/ts/idilia/DocMatchingServiceTest.java
@@ -32,9 +32,9 @@ public class DocMatchingServiceTest {
     tweets.add(new Tweet(1, "tide"));
     tweets.add(new Tweet(2, "The tide was high when he was at the beach"));
     List<Sense> senses = Collections.singletonList(new Sense(0, 1, "tide", "Tide/N8"));
-    List<Integer> docRes = matchingSvc.matchAsync(senses, tweets, customerId).join();
-    Integer docResA[] = new Integer[3];
-    assertArrayEquals(new Integer[]{1,0,-1}, docRes.toArray(docResA));
+    List<Double> docRes = matchingSvc.matchAsync(senses, tweets, customerId).join();
+    Double docResA[] = new Double[3];
+    assertArrayEquals(new Double[]{1.0,0.0,-1.0}, docRes.toArray(docResA));
   }
